Remove --domains option
It was confusing as hell, it even confused me LOL, because
when used in a command line according to the help text,
it didn't split by white space, only when using the envvar.

It was only used to get the domains from the CLOUDFLARE_DOMAINS
environment variable, but they are now read directly from os.environ
instead of parsing them with click, the same way as before.