Added pydantic dependency
It will be needed for the cache structure, because I want to switch
to JSON as cache storage instead of Pickle for longevity and
backward-compatibility, but attrs is not very good at JSON serialization
for even simple types like ipaddress.
Pydantic is only compatible with Python3.6.1+, so required Python version
had to be bumped.