Dynamically calculate zone from domains
So far we assumed that zone name is the last two parts of the domain,
but this is not true, for example example.co.uk is a valid zone name.

Get the list of all zones, compare the domain to all of them and get the
zone name which is at the end of the domain.

Fixes #35.