Fix exit code never propagated
Not sure when this went wrong, but the default of a Click application
is standalone_mode=True, which means the main function never returns,
exceptions are caught and the exit code will be set based on the
exc.exit_code property.

hatchling will wrap the main function simply with sys.exit(main()),
but as the return value was never returned by Click, this way always 0.

Fixes #42